refactor: Удален вспомогательный класс Itenirary
В нем хранились методы сортровки и вспомогательные функции для
отображений. Методы сортировки переехали в модель SortingsModel,
функции — в отображения EventSummaryView и EventFormView.
